# swagger/api.yaml
swagger: '2.0'
info:
  description: Documentation for interacting with Simple Riqum via RESTful/HTTP
  version: 1.0.0
  title: Simple Riqum
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: giakhanh2487@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8080'
basePath: /api/v1
tags:
  - name: auth
    description: Authorization
  - name: user
    description: Everything about your Users
  - name: resource
    description: Everything about your Resources
  - name: quota
    description: Everything about your Quota
schemes:
  - http
  - https
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Authenticate user
      description: ''
      operationId: login
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Login credential
          required: true
          schema:
            $ref: '#/definitions/BasicCredential'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/LoginResponse'
        '400':
          description: invalid login request
          schema:
            $ref: '#/definitions/ApiResponse'

  /auth/logout:
    post:
      tags:
        - auth
      summary: Logout a user
      description: ''
      operationId: logout
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        default:
          description: successful operation
  /users:
    post:
      tags:
        - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      security:
        - BasicAuth: []
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: User has been created
          schema:
            $ref: '#/definitions/User'
        '400':
          description: invalid post request
          schema:
            $ref: '#/definitions/ApiResponse'
        '403':
          schema:
            $ref: '#/definitions/ApiResponse'
    get:
      tags:
        - user
      security:
        - BasicAuth: []
      summary: Get one or more users
      description: ''
      operationId: findUsers
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          type: integer
          description: Page number to retrieve
        - name: size
          in: query
          type: integer
          description: Elements per page to retrieve
      responses:
        '200':
          description: A list of users
          schema:
            type: array
            items:
              $ref: '#/definitions/User'

        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'

        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'

  '/users/me':
    get:
      tags:
      - user
      security:
      - BasicAuth: []
      summary: Get one user identified by current token
      description: ''
      operationId: getUser
      produces:
      - application/json

      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid Request
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/ApiResponse'

  '/users/{userID}':
    get:
      tags:
        - user
      security:
        - BasicAuth: []
      summary: Get one user identified by userID
      description: ''
      operationId: getUser
      produces:
        - application/json
      parameters:
        - name: userID
          in: path
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid Request
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - user
      security:
        - BasicAuth: []
      summary: Update one user identified by userID
      description: ''
      operationId: updateUser
      produces:
        - application/json
      parameters:
        - name: userID
          in: path
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Created user object
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              roles:
                type: array
                items:
                  type: string

      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid input values
    delete:
      tags:
      - user
      security:
      - BasicAuth: []
      summary: Remove one user identified by userID
      description: ''
      operationId: deleteUser
      parameters:
      - name: userID
        in: path
        required: true
        type: integer
        format: int64
      responses:
        '204':
          description: User has been successfully removed
        '400':
          description: Invalid Request
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/ApiResponse'

  '/users/{userID}/quota':
    get:
      tags:
      - user
      security:
      - BasicAuth: []
      summary: Update quota for user identified by userID
      description: ''
      operationId: getUser
      produces:
      - application/json
      parameters:
      - name: userID
        in: path
        required: true
        type: integer
        format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Quota'
        '400':
          description: Invalid Request
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
      - user
      security:
      - BasicAuth: []
      summary: Update one user identified by userID
      description: ''
      operationId: updateUser
      produces:
      - application/json
      parameters:
      - name: userID
        in: path
        required: true
        type: integer
        format: int64
      - in: body
        name: body
        description: New quota to update
        required: true
        schema:
          $ref: '#/definitions/Quota'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Quota'
        '400':
          description: Invalid Request
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/ApiResponse'
  '/resources':
    post:
      tags:
        - resource
      security:
        - BasicAuth: []
      summary: Create new resource
      description: By default, new resource belongs to the current user
      parameters:
      - in: body
        name: body
        description: Created resource object
        required: true
        schema:
          $ref: '#/definitions/Resource'
      responses:
        '201':
          description: successful operation
          schema:
            $ref: '#/definitions/Resource'
        '400':
          description: invalid post request
          schema:
            $ref: '#/definitions/ApiResponse'
        '403':
          schema:
            $ref: '#/definitions/ApiResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
    get:
      tags:
        - resource
      security:
      - BasicAuth: []
      summary: Get one or more resources
      description: By default, only resources under current user's ownership are retrieved
      parameters:
      - name: page
        in: query
        type: integer
        description: Page number to retrieve
      - name: size
        in: query
        type: integer
        description: Elements per page to retrieve
      responses:
        '200':
          description: A list of resources
          schema:
            type: array
            items:
              $ref: '#/definitions/Resource'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
  '/resources/{resourceId}':
    get:
      tags:
        - resource
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: resourceId
          type: string
          required: true
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Resource'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - resource
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: resourceId
          type: string
          required: true
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Resource'
        '400':
          description: Invalid input resourceId
    delete:
      tags:
        - resource
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: resourceId
          type: string
          required: true
      responses:
        '204':
          description: empty content
        '400':
          description: Invalid resourceId
      
securityDefinitions:
  BasicAuth:
    type: basic
  OAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: 'https://riqum.khanhhua.com/oauth/authorize'
    tokenUrl: 'https://riqum.khanhhua.com/oauth/token'
    scopes:
      user: Grants read/write to his own resources
      admin: >-
        Grants read and write access to administrative information (user, grant
        to users)

definitions:
  BasicCredential:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
      password:
        type: string
  User:
    type: object
    required:
      - username
      - email
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      userStatus:
        type: integer
        format: int64
        description: User Status
      roles:
        type: array
        items:
          type: string
      quotum:
        schema:
          $ref: '#/definitions/Quota'
    xml:
      name: User
  Quota:
    type: object
    required:
      - limit
      - unit
    properties:
      limit:
        type: integer
      usage:
        type: integer
      unit:
        type: string

  Resource:
    type: object
    required:
      - name
    properties:
      id:
        type: string
      name:
        type: string
      ownerId:
        type: integer
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Category

  LoginResponse:
    type: object
    properties:
      accessToken:
        type: string

  ApiResponse:
    type: object
    properties:
      code:
        type: integer
        format: int64
      type:
        type: string
      message:
        type: string
      errors:
        type: array
        items:
          type: string
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'